Added a SyntaxUnit structure and changed existing functions to generate a syntax tree instead of parse tree
The syntax unit is an extension of a token unit with some addition information, ScopeType as context,
this will help to disambiguate tree branches when reading and searching so that the context of the branch and node is explicit rather than implicit based on the position of the branch in relation to some node

I suspect there are some bugs in getting the appropriate the context for each unit, even now, tokens outside of brackets are often assigned the wrong context
but for cases like function definition, these can be ignored. Some more experimentation will have to be done to determine if this is a dangerous bug or not.

There is still some positional dependence in the tree, for example, when defining a function, the name of the function should always be the first branch of the keyword node, but there is no explicit data in the syntax unit to denote a 'name' for a function under a declaration.
This must be inferred based on node and branch position. Not necessarily ideal but it is fair to ask of the developer that a function name is always required and should always be immediately following a declaration keyword