The lack of tokenizing negative numbers was an embarrassing oversight but it should be working properly now.
Had to hack together a solution in lexer where the addSpaces function uses O.spacingRepr instead of O.repr to addspaces to operators
This will introduce some bugs and inconsistencies.

Hopefully, if I can find a way to add some robust error-checking, the addSpaces function may become unnecessary in the face of having a useful grammar linter or something.