removed references to ExecEnv, going to test to see if any functionality was unintentionally disrupted
The only test affected, aside from some hacky bug fixes I put together yesterday,

is that later value bindings do no override earlier value bindings of the same name
so I suspect that is because there is something not stack-like here or that the stack get's built in the reverse order it was before.
Either way, I don't think it's the end of the world and I will stick with this.