Made some changes to the way Sakana will execute a program.
Now in the main scope, the 'swim' keyword is used to denote the main function.
procedural programming can be done with send fish and send fish can be used to bind values to scopes below them.
You no longer need nested fish to concurrently execute IO statements and a fish program will return whatever is the value of the first return fish in a swim block