Introducing the abstract Keyword "Lamprey"
This keyword is used in the parser and interpreter to hold a function under one tree, even if the function is not bound to an id. This is so that ALL functions, bound or unbound can be represented as one tree and not a minimum of two trees for arguments and returns.
But the lamprey node holds the arguments and return under itself.

# Conflicts:
#	Test/SyntaxTreeTest.hs
#	app/SakanaParser.hs